% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtreg-package.R
\docType{package}
\name{dtreg-package}
\alias{dtreg}
\alias{dtreg-package}
\title{dtreg: Interact with Data Type Registries and Create Machine-Readable Data}
\description{
You can load a schema from a DTR (data type registry) as an R object. Use this schema to write your data in JSON-LD (JavaScript Object Notation for Linked Data) format to make it machine readable.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/TIBHannover/orkg/dtreg-r}
  \item Report bugs at \url{https://gitlab.com/TIBHannover/orkg/dtreg-r/-/issues}
}

}
\author{
\strong{Maintainer}: Olga Lezhnina \email{olga.lezhnina@tib.eu} (\href{https://orcid.org/0000-0003-2237-7725}{ORCID})

Authors:
\itemize{
  \item Manuel Prinz \email{manuel.prinz@tib.eu} (\href{https://orcid.org/0000-0003-2151-4556}{ORCID})
  \item Markus Stocker \email{markus.stocker@tib.eu} (\href{https://orcid.org/0000-0001-5492-3212}{ORCID})
}

Other contributors:
\itemize{
  \item Open Research Knowledge Graph Project and Contributors [copyright holder]
}

}
\keyword{internal}
