% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{show_fields}
\alias{show_fields}
\title{Show schema fields}
\usage{
show_fields(datatype)
}
\arguments{
\item{datatype}{An R6 class from dtreg::load_datatype}
}
\value{
A character vector of available fields
}
\description{
Show which fields of a schema can be used for creating an instance
}
\examples{
dt <- dtreg::load_datatype("https://doi.org/21.T11969/aff130c76e68ead3862e")
dtreg::show_fields(dt$data_item())

}
